Fixed: save reputation standing as signed value. This will prevent DB erro at overflow value in `standing` field.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@2923 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Player.cpp
@@ -9885,7 +9885,7 @@ void Player::_LoadReputation()
 
             newFaction.ID               = fields[0].GetUInt32();
             newFaction.ReputationListID = fields[1].GetUInt32();
-            newFaction.Standing         = fields[2].GetUInt32();
+            newFaction.Standing         = int32(fields[2].GetUInt32());
             newFaction.Flags            = fields[3].GetUInt32();
 
             factions.push_back(newFaction);
@@ -10290,7 +10290,7 @@ void Player::_SaveReputation()
 
     for(itr = factions.begin(); itr != factions.end(); ++itr)
     {
-        sDatabase.PExecute("INSERT INTO `character_reputation` (`guid`,`faction`,`reputation`,`standing`,`flags`) VALUES ('%u', '%u', '%u', '%u', '%u')", GetGUIDLow(), itr->ID, itr->ReputationListID, itr->Standing, itr->Flags);
+        sDatabase.PExecute("INSERT INTO `character_reputation` (`guid`,`faction`,`reputation`,`standing`,`flags`) VALUES ('%u', '%u', '%u', '%i', '%u')", GetGUIDLow(), itr->ID, itr->ReputationListID, itr->Standing, itr->Flags);
     }
     sDatabase.CommitTransaction();
 }
